requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.11.16; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
altair==5.5.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
faiss-cpu==1.10.0; python_version >= '3.9'
filelock==3.18.0; python_version >= '3.9'
frozenlist==1.5.0; python_version >= '3.8'
fsspec==2025.3.2; python_version >= '3.9'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.44; python_version >= '3.7'
greenlet==3.1.1; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
huggingface-hub==0.30.1; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
langchain==0.3.23; python_version >= '3.9' and python_version < '4.0'
langchain-community==0.3.21; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.51; python_version >= '3.9' and python_version < '4.0'
langchain-huggingface==0.1.2; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.8; python_version >= '3.9' and python_version < '4.0'
langsmith==0.3.24; python_version >= '3.9' and python_version < '4.0'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
mpmath==1.3.0
multidict==6.3.2; python_version >= '3.9'
mypy-extensions==1.0.0; python_version >= '3.5'
narwhals==1.33.0; python_version >= '3.8'
networkx==3.4.2; python_version >= '3.10'
numpy==2.2.4; python_version >= '3.10'
orjson==3.10.16; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.1.0; python_version >= '3.9'
propcache==0.3.1; python_version >= '3.9'
protobuf==5.29.4; python_version >= '3.8'
pyarrow==19.0.1; python_version >= '3.9'
pydantic==2.11.2; python_version >= '3.9'
pydantic-core==2.33.1; python_version >= '3.9'
pydantic-settings==2.8.1; python_version >= '3.8'
pydeck==0.9.1; python_version >= '3.8'
pypdf==5.4.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.0; python_version >= '3.9'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rpds-py==0.24.0; python_version >= '3.9'
safetensors==0.5.3; python_version >= '3.7'
scikit-learn==1.6.1; python_version >= '3.9'
scipy==1.15.2; python_version >= '3.10'
sentence-transformers==4.0.2; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.2; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.40; python_version >= '3.7'
streamlit==1.44.1; python_version >= '3.9' and python_full_version != '3.9.7'
sympy==1.13.1; python_version >= '3.8'
tenacity==9.1.2; python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
tokenizers==0.21.1; python_version >= '3.9'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
torch==2.6.0; python_full_version >= '3.9.0'
tornado==6.4.2; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
transformers==4.50.3; python_full_version >= '3.9.0'
typing-extensions==4.13.1; python_version >= '3.8'
typing-inspect==0.9.0
typing-inspection==0.4.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
watchdog==6.0.0; python_version >= '3.9'
yarl==1.18.3; python_version >= '3.9'
zstandard==0.23.0; python_version >= '3.8'
